Update Statistic services

+++ RGBA.Optio.Domain/Models/ResponseModels/CategoryResponseModel.cs
@@ -5,7 +5,7 @@ namespace RGBA.Optio.Domain.Models.ResponseModels
     {
         public  string TransactionCategory { get; set; }
         public  long TransactionTypeID { get; set; }
-        public long TransactionCount {  get; set; }
-        public long TransactionVolume {  get; set; }
+        public decimal TransactionCount {  get; set; }
+        public decimal TransactionVolume {  get; set; }
     }
 }
